wr-231: * changed ConfirmDialog to ConfirmArticleDeleteDialog
