+++ frontend/src/libs/components/components.ts
@@ -3,7 +3,7 @@ export { Avatar } from './avatar/avatar.js';
 export { Button } from './button/button.js';
 export { CommentCard } from './comment-card/comment-card.js';
 export { Comments } from './comments/comments.js';
-export { ConfirmDialog } from './confirm-dialog/confirm-dialog.js';
+export { ConfirmArticleDeleteDialog } from './confirm-article-delete-dialog/confirm-article-delete-dialog.js';
 export { ErrorMessage } from './error-message/error-message.js';
 export { Header } from './header/header.js';
 export { Icon } from './icon/icon.js';
