+++ frontend/src/pages/article/components/article-view/styles.module.scss
@@ -103,16 +103,6 @@
   word-break: break-word;
 }
 
-.buttonDelete {
-  color: var(--white);
-  background-color: var(--button-delete-bg);
-}
-
-.buttonDelete:hover {
-  color: var(--white);
-  background-color: var(--button-delete-hover);
-}
-
 .presentationAuthorName {
   display: none;
   padding: 20px 33px 0;
