+++ frontend/src/pages/articles/components/article-card/styles.module.scss
@@ -228,13 +228,3 @@
 .moreActionsIcon:hover {
   background-color: var(--light-gray);
 }
-
-.buttonDelete {
-  color: var(--white);
-  background-color: var(--button-delete-bg);
-}
-
-.buttonDelete:hover {
-  color: var(--white);
-  background-color: var(--button-delete-hover);
-}
